Fix sub-package tests
