--- python/tests/test_adnormal.py
@@ -1,12 +1,0 @@
-import numpy as np
-from numpy.testing import assert_allclose
-
-from light_curve_pure import AndersonDarlingNormal
-
-
-def test_adnormal():
-    m = np.arange(0, 9)
-    feature = AndersonDarlingNormal()
-    desired = feature(m, m, None)
-    actual = 0.155339690
-    assert_allclose(actual, desired)
