--- python/tests/test_medabsdev.py
@@ -1,11 +1,0 @@
-from numpy.testing import assert_allclose
-
-from light_curve_pure import MedianAbsoluteDeviation
-
-
-def test_medabsdev():
-    m = [1.0, 1.0, 3.0, 4.0, 7.0]
-    feature = MedianAbsoluteDeviation()
-    actual = feature(m, m, None)
-    desired = 2.0
-    assert_allclose(actual, desired)
